# draft: measure e2e latency for amazon q requests