--- a/packages/core/src/codewhispererChat/view/connector/connector.ts
+++ b/packages/core/src/codewhispererChat/view/connector/connector.ts
@@ -141,6 +141,7 @@ export interface ChatMessageProps {
     readonly codeReference?: CodeReference[]
     readonly triggerID: string
     readonly messageID: string
+    readonly traceId?: string
 }
 
 export class ChatMessage extends UiMessage {
@@ -153,6 +154,7 @@ export class ChatMessage extends UiMessage {
     readonly followUpsHeader: string | undefined
     readonly triggerID: string
     readonly messageID: string | undefined
+    readonly traceId?: string
     override type = 'chatMessage'
 
     constructor(props: ChatMessageProps, tabID: string) {
@@ -165,6 +167,7 @@ export class ChatMessage extends UiMessage {
         this.codeReference = props.codeReference
         this.triggerID = props.triggerID
         this.messageID = props.messageID
+        this.traceId = props.traceId
     }
 }